backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2  # For Excel file support
python-dateutil==2.8.2

# Validation & Type Checking
pydantic==2.5.3
pydantic-settings==2.1.0

# Statistics & Algorithms
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Optional: For fuzzy matching (Phase 2)
# fuzzywuzzy==0.18.0
# python-Levenshtein==0.23.0